Mokito Exercise separated from JUnit with Updated Documentation
