--- Week 2/JUnit/junit_exercise/src/main/java/com/example/ExternalApi.java
@@ -1,5 +1,0 @@
-package com.example;
-
-public interface ExternalApi {
-    String getData();
-}
