--- Week 2/JUnit/junit_exercise/src/main/java/com/example/MyService.java
@@ -1,14 +1,0 @@
-package com.example;
-
-public class MyService {
-
-    private final ExternalApi externalApi;
-
-    public MyService(ExternalApi externalApi) {
-        this.externalApi = externalApi;
-    }
-
-    public String fetchData() {
-        return externalApi.getData();
-    }
-}
