--- Week 2/JUnit/junit_exercise/src/test/java/com/example/MyServiceTest.java
@@ -1,31 +1,0 @@
-package com.example;
-
-import static org.junit.Assert.assertEquals;
-import org.junit.Test;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-public class MyServiceTest {
-
-    @Test
-    public void testExternalApi() {
-        // Step 1: Create a mock of ExternalApi
-        ExternalApi mockApi = mock(ExternalApi.class);
-
-        // Step 2: Stub getData() to return a fake value
-        when(mockApi.getData()).thenReturn("Mock Data");
-
-        // Step 3: Inject mock into MyService
-        MyService service = new MyService(mockApi);
-
-        // Step 4: Call method under test
-        String result = service.fetchData();
-
-        // Step 5: Assert the result
-        assertEquals("Mock Data", result);
-
-        // Optional: Verify the interaction
-        verify(mockApi).getData();
-    }
-}
