+++ Week 2/JUnit_Mokito/mockito_exercise/src/main/java/com/cognizant/App.java
@@ -1,0 +1,13 @@
+package com.cognizant;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
