+++ Week 2/JUnit_Mokito/mockito_exercise/src/main/java/com/cognizant/ExternalApi.java
@@ -1,0 +1,5 @@
+package com.cognizant;
+
+public interface ExternalApi {
+    String getData();
+}
