+++ Week 2/JUnit_Mokito/mockito_exercise/src/main/java/com/cognizant/MyService.java
@@ -1,0 +1,14 @@
+package com.cognizant;
+
+public class MyService {
+
+    private final ExternalApi externalApi;
+
+    public MyService(ExternalApi externalApi) {
+        this.externalApi = externalApi;
+    }
+
+    public String fetchData() {
+        return externalApi.getData();
+    }
+}
