+++ Week 2/JUnit_Mokito/mockito_exercise/src/test/java/com/cognizant/AppTest.java
@@ -1,0 +1,12 @@
+package com.cognizant;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.junit.jupiter.api.Test;
+
+public class AppTest {
+
+    @Test
+    void testSomething() {
+        assertTrue(true);
+    }
+}
