+++ Week 2/SLF4J_Logging/slf4j_exercise/src/main/java/com/cognizant/LoggingExample.java
@@ -1,3 +1,5 @@
+package com.cognizant;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
